+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   will be unable to use SpellBot in channels that require verification.
 - Adds a `!verify` command that can be used to verify users mentioned.
 - Adds an `!unverify` command that does the opposite of `!verify`.
+- Adds a `!spellbot cmotd` command to set the channel message of the day.
 
 ### Changed